Float -> double.

+++ jigsaw/Drag.h
@@ -9,9 +9,9 @@ public:
 
 private:
 	int selected_x, selected_y;
-    float offset_x, offset_y;
-    float drag_offset_x, drag_offset_y;
-    float drag_end_x, drag_end_y;
+    double offset_x, offset_y;
+    double drag_offset_x, drag_offset_y;
+    double drag_end_x, drag_end_y;
     int rectsizex, rectsizey;
     bool leftdrag, rightdrag, rect_up;
 };
