+++ jigsaw/Piece.h
@@ -9,12 +9,12 @@ public:
     Piece(int x, int y, int nx, int ny);
     virtual ~Piece();
     void Render();
-    void CustomRender(float x, float y);
+    void CustomRender(double x, double y);
     void Move(int x, int y);
     bool Matches(int x, int y);
 
 	void set_color(Color c) { color = c; }
-	void set_alpha(float a) { alpha = a; }
+	void set_alpha(double a) { alpha = a; }
 	void QueueRenderPosition(Vector2D pos) {
 		queue_pos = pos;
 		position_queued = true;
@@ -23,7 +23,7 @@ private:
 	Vector2D position, size, tex_origin, queue_pos;
 	bool position_queued;
 	Color color;
-	float alpha;
+	double alpha;
     void InternalRender();
 };
